Refactor the "git fat checkout" operation (and therefore also "git fat pull") to significantly improve its speed.
The main change here is that git checkout-index is now called only a single time for all files.
In some tests the performance improvement is six-fold.

With this change it doesn't make sense anymore to print out each file as it is being processed,
so only the number of files that need to be restored is shown.
The actual filenames are only printed if git-fat is verbose.